Kiiiinda forgot I had this in a repo
So far I've set up a rough frame-work for the entire app and a database. I've mainly been focusing on the API integration first, and had to switch APIs after working with the first one for a while. The recipe search page is nearing completion; it returns 10 results. I will add an advanced filter panel if I get time, as well as pagination if possible (I can't actually see a way to return more than 10 results). You can click on any of the results to see the full recipe, and I just need to write the function for the 'save to My Recipes' button once I've started messing with my db