functionally the search recipes page is now done
I've also turned the recipe viewing portion of this into it's own component and added that into the My Recipes page.  This means the My Recipes page works exactly like the recipe search page, but without the search.  The recipes displayed are those that have been saved by the user, and from here the 'save' button is now a 'remove' button.